queue: split out queue helper structs
Split out AvailIter, Descriptor, DescriptorChain to dedicated files.
There should no visible functional changes.
